+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.2.4 (unreleased)
 
 - Updated NGT to 1.9.1
+- Added support for passing an index to optimizers
 
 ## 0.2.3 (2020-03-08)